move devnet fixtures to the new file
