+++ starknet_py/conftest.py
@@ -4,7 +4,8 @@ pytest_plugins = [
     "starknet_py.tests.e2e.fixtures.clients",
     "starknet_py.tests.e2e.fixtures.account_clients",
     "starknet_py.tests.e2e.fixtures.contracts",
-    "starknet_py.tests.e2e.fixtures.utils",
+    "starknet_py.tests.e2e.fixtures.misc",
+    "starknet_py.tests.e2e.fixtures.devnet",
     "starknet_py.tests.e2e.fixtures.constants",
     "starknet_py.tests.e2e.client.fixtures.transactions",
     "starknet_py.tests.e2e.client.fixtures.prepare_network",
