+++ starknet_py/tests/e2e/fixtures/misc.py
@@ -1,12 +1,6 @@
 # pylint: disable=redefined-outer-name
 
 import json
-import subprocess
-import time
-from contextlib import closing
-import socket
-from typing import Generator
-
 import pytest
 import pytest_asyncio
 
@@ -43,45 +37,6 @@ def pytest_collection_modifyitems(config, items):
         )
         if should_not_run:
             item.add_marker(pytest.mark.skip())
-
-
-def get_available_port() -> int:
-    with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as sock:
-        sock.bind(("", 0))
-        sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        return sock.getsockname()[1]
-
-
-def start_devnet():
-    devnet_port = get_available_port()
-
-    command = [
-        "poetry",
-        "run",
-        "starknet-devnet",
-        "--host",
-        "localhost",
-        "--port",
-        str(devnet_port),
-        "--accounts",  # deploys specified number of accounts
-        str(1),
-        "--seed",  # generates same accounts each time
-        str(1),
-    ]
-    # pylint: disable=consider-using-with
-    proc = subprocess.Popen(command)
-    time.sleep(5)
-    return devnet_port, proc
-
-
-@pytest.fixture(scope="module")
-def run_devnet() -> Generator[str, None, None]:
-    """
-    Runs devnet instance once per module and returns it's address
-    """
-    devnet_port, proc = start_devnet()
-    yield f"http://localhost:{devnet_port}"
-    proc.kill()
 
 
 @pytest.fixture(scope="module")
